% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_functions.R
\name{load_pangenomeRDS}
\alias{load_pangenomeRDS}
\title{Load A Pagoo Pangenome}
\usage{
load_pangenomeRDS(file, seqs.if.avail = TRUE)
}
\arguments{
\item{file}{The path to the pangenome `.RDS` file.}

\item{seqs.if.avail}{\code{logical} Whether to load sequences or not, if they
are available in the rds file.}
}
\value{
Ethier a \code{PgR6MS} class object, or a \code{PgR6M} object (with or
without sequences, respectively).
}
\description{
This function loads a pagoo pangenome from a `.RDS` file generated
by the `save_pangenomeRDS` method. Objects loaded by this functions keep thier
states, i.e : dropped/recovered organisms are conserved, as well as the
`core_level` setted when the object was originally saved.
}
