% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PgR6.R
\docType{data}
\name{PgR6}
\alias{PgR6}
\title{PgR6 basic class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PgR6
}
\description{
A basic \code{pgr6} class constructor.
}
\keyword{datasets}
