% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PgR6MS.R
\docType{data}
\name{PgR6MS}
\alias{PgR6MS}
\title{Expect 1) a named list of named character vector Names of list is names of
organisms, names of character vector are gene names; 2) a named list of
BStringSet objects (same requirements as (1), but with BStringSet names as
gene names); 3) a BStringSetList (same requirements as (3) but
BStringSetList names are organisms names).}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PgR6MS
}
\description{
Expect 1) a named list of named character vector Names of list is names of
organisms, names of character vector are gene names; 2) a named list of
BStringSet objects (same requirements as (1), but with BStringSet names as
gene names); 3) a BStringSetList (same requirements as (3) but
BStringSetList names are organisms names).
}
\keyword{datasets}
